packfile.h: move pack_window and pack_entry from cache.h
